Fix basic UI errors
get_visible_dialogs, store_dimensions and preferences_dialog_init
now works
